+++ packages/local-environment/bootup.sh
@@ -1,5 +1,6 @@
 #!/bin/bash
+export $(cat ./.env | xargs)
 pushd ./hardhat
-docker build -t hardhat:latest . 
+docker build -t hardhat:latest .
 popd
-docker pull gcr.io/nomad-xyz/nomad-agent:prestwich-remove-deploy-gas
+docker pull "$AGENTS_IMAGE"